Compiles & runs in windows (64bit and 32bit)

+++ src/include/robin_hood.h
@@ -415,7 +415,7 @@ public:
 		h *= m;
 		h ^= h >> r;
 
-		return h;
+		return static_cast<size_t>(h);
 	}
 };
